BackEnd/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
Pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
opcua==0.98.13
cryptography==45.0.5
